refactor(dashboard): AI suggestions dialog and introduce new components
- Consolidated AI suggestions logic by creating dedicated components for initial, loading, and error states.
- Introduced a constants file for prompt suggestions categorized by step type, enhancing maintainability.
- Added utility functions for parsing variables and managing nested values, improving code clarity.
- Implemented a content preview component to render AI-generated content dynamically based on step type.
- Enhanced the overall structure of the AI suggestions dialog for better readability and performance.
